Add loveholidays as using vite
The lower case `l` is deliberate.

We're a UK based package holiday company with a customer base around Europe. 

We're using vite on our next gen frontend. Currently this doesn't handle much of our site, but we're building it up. 

As listed we're using Preact and vitest. Some other interesting technologies we're using on the project are urql, turborepo, typescript, pandacss.

We're a big fan of vite's build performance, good conventions around dynamic imports and our ability to write plugins to solve a couple problems specific to us.